+++ sdk/communication/Azure.Communication.CallAutomation/src/CallAutomationClientOptions.cs
@@ -16,7 +16,7 @@ namespace Azure.Communication.CallAutomation
         /// <summary>
         /// The latest version of the CallAutomation service.
         /// </summary>
-        internal const ServiceVersion LatestVersion = ServiceVersion.V2023_10_15;
+        internal const ServiceVersion LatestVersion = ServiceVersion.V2024_01_22_Preview;
 
         internal string ApiVersion { get; }
 
@@ -34,6 +34,7 @@ namespace Azure.Communication.CallAutomation
             {
                 ServiceVersion.V2023_03_06 => "2023-03-06",
                 ServiceVersion.V2023_10_15 => "2023-10-15",
+                ServiceVersion.V2024_01_22_Preview => "2024-01-22-preview",
                 _ => throw new ArgumentOutOfRangeException(nameof(version)),
             };
         }
@@ -50,9 +51,15 @@ namespace Azure.Communication.CallAutomation
             V2023_03_06 = 1,
 
             /// <summary>
-            /// The GA2 of the CallAutomation service.
+            /// The GA1 of the CallAutomation service.
             /// </summary>
-            V2023_10_15 = 2
+#pragma warning disable CA1707 // Identifiers should not contain underscores
+            V2023_10_15 = 2,
+
+            /// <summary>
+            /// The BETA3 of the CallAutomation service.
+            /// </summary>
+            V2024_01_22_Preview = 3
 #pragma warning restore CA1707 // Identifiers should not contain underscores
         }
     }