+++ sdk/communication/Azure.Communication.CallAutomation/tests/EventProcessors/ResultWithEventProcessorAsyncTests.cs
@@ -581,11 +581,87 @@ namespace Azure.Communication.CallAutomation.Tests.EventProcessors
             // Assert
             Assert.NotNull(returnedResult);
             Assert.AreEqual(true, returnedResult.IsSuccess);
+
             Assert.NotNull(returnedResult.SuccessResult);
             Assert.IsNull(returnedResult.FailureResult);
             Assert.AreEqual(typeof(CancelAddParticipantSucceeded), returnedResult.SuccessResult.GetType());
             Assert.AreEqual(CallConnectionId, returnedResult.SuccessResult.CallConnectionId);
             Assert.AreEqual(invitationId, returnedResult.SuccessResult.InvitationId);
+        }
+
+        public async Task StartDialogEventResultSuccessTest()
+        {
+            int successCode = (int)HttpStatusCode.Created;
+
+            var callConnection = CreateMockCallConnection(successCode, DialogPayload);
+            CallAutomationEventProcessor handler = callConnection.EventProcessor;
+
+            var dialogContext = new Dictionary<string, object>();
+            var startDialogOptions = new StartDialog(new PowerVirtualAgentsDialog("botAppId", dialogContext))
+            {
+                OperationContext = OperationContext
+            };
+
+            var response = callConnection.GetCallDialog().StartDialog(startDialogOptions);
+            Assert.AreEqual(successCode, response.GetRawResponse().Status);
+
+            // Create and send event to event processor
+            SendAndProcessEvent(handler, new DialogStarted(CallConnectionId, ServerCallId, CorelationId, OperationContext, new ResultInformation(), "dialogId", DialogInputType.PowerVirtualAgents));
+
+            DialogEventResult returnedResult = await response.Value.WaitForEventProcessorAsync();
+
+            // Assert
+            Assert.NotNull(returnedResult);
+            Assert.AreEqual(true, returnedResult.IsSuccess);
+            Assert.NotNull(returnedResult.DialogStartedSuccessEvent);
+            Assert.IsNull(returnedResult.DialogCompletedSuccessResult);
+            Assert.IsNull(returnedResult.DialogConsentSuccessEvent);
+            Assert.IsNull(returnedResult.DialogHangupSuccessEvent);
+            Assert.IsNull(returnedResult.DialogLanguageChangeEvent);
+            Assert.IsNull(returnedResult.DialogSensitivityUpdateEvent);
+            Assert.IsNull(returnedResult.DialogTransferSuccessEvent);
+            Assert.IsNull(returnedResult.FailureResult);
+            Assert.AreEqual(typeof(DialogStarted), returnedResult.DialogStartedSuccessEvent.GetType());
+            Assert.AreEqual(CallConnectionId, returnedResult.DialogStartedSuccessEvent.CallConnectionId);
+            Assert.AreEqual(OperationContext, returnedResult.DialogStartedSuccessEvent.OperationContext);
+        }
+
+        [Test]
+        public async Task StartDialogEventResultFailedTest()
+        {
+            int successCode = (int)HttpStatusCode.Created;
+
+            var callConnection = CreateMockCallConnection(successCode, DialogPayload);
+            CallAutomationEventProcessor handler = callConnection.EventProcessor;
+
+            var dialogContext = new Dictionary<string, object>();
+            var startDialogOptions = new StartDialog(new PowerVirtualAgentsDialog("botAppId", dialogContext))
+            {
+                OperationContext = OperationContext
+            };
+
+            var response = callConnection.GetCallDialog().StartDialog(startDialogOptions);
+            Assert.AreEqual(successCode, response.GetRawResponse().Status);
+
+            // Create and send event to event processor
+            SendAndProcessEvent(handler, new DialogFailed(CallConnectionId, ServerCallId, CorelationId, OperationContext, new ResultInformation(), "dialogId", DialogInputType.PowerVirtualAgents));
+
+            DialogEventResult returnedResult = await response.Value.WaitForEventProcessorAsync();
+
+            // Assert
+            Assert.NotNull(returnedResult);
+            Assert.AreEqual(false, returnedResult.IsSuccess);
+            Assert.IsNull(returnedResult.DialogStartedSuccessEvent);
+            Assert.IsNull(returnedResult.DialogCompletedSuccessResult);
+            Assert.IsNull(returnedResult.DialogConsentSuccessEvent);
+            Assert.IsNull(returnedResult.DialogHangupSuccessEvent);
+            Assert.IsNull(returnedResult.DialogLanguageChangeEvent);
+            Assert.IsNull(returnedResult.DialogSensitivityUpdateEvent);
+            Assert.IsNull(returnedResult.DialogTransferSuccessEvent);
+            Assert.NotNull(returnedResult.FailureResult);
+            Assert.AreEqual(typeof(DialogFailed), returnedResult.FailureResult.GetType());
+            Assert.AreEqual(CallConnectionId, returnedResult.FailureResult.CallConnectionId);
+            Assert.AreEqual(OperationContext, returnedResult.FailureResult.OperationContext);
         }
     }
 }