+++ sdk/communication/Azure.Communication.CallAutomation/tests/Infrastructure/CallAutomationClientAutomatedLiveTestsBase.cs
@@ -98,7 +98,7 @@ namespace Azure.Communication.CallAutomation.Tests.Infrastructure
         {
             var connectionString = TestEnvironment.LiveTestStaticConnectionString;
 
-            CallAutomationClient callAutomationClient = new CallAutomationClient(connectionString, CreateServerCallingClientOptionsWithCorrelationVectorLogs(source));
+            CallAutomationClient callAutomationClient = new CallAutomationClient(new Uri("https://x-pma-uswe-02.plat.skype.com:6448/"), connectionString, CreateServerCallingClientOptionsWithCorrelationVectorLogs(source));
 
             return InstrumentClient(callAutomationClient);
         }