+++ sdk/communication/Azure.Communication.CallAutomation/tests/Infrastructure/CallAutomationClientTestEnvironment.cs
@@ -17,6 +17,8 @@ namespace Azure.Communication.CallAutomation.Tests.Infrastructure
 
         public const string TargetNumber = "TARGET_PHONE_NUMBER";
 
+        public const string Endpoint = "PMA_Endpoint";
+
         private const string randomResourceIdentifier = "82e890fc-188a-4b67-bb7d-deff073d7d1e";
 
         private string randomAcsUser = $"8:acs:{randomResourceIdentifier}_0000000e-abbe-44ad-9f37-b0a72a616d0b";
@@ -26,6 +28,8 @@ namespace Azure.Communication.CallAutomation.Tests.Infrastructure
         private string dispatcherEndpoint = "DISPATCHER_ENDPOINT";
 
         private string servicebusString = "SERVICEBUS_STRING";
+
+        private string botAppId = "BOT_APP_ID";
 
         /// <summary>
         /// The resource identifier associated with the Azure Communication Service.
@@ -52,9 +56,19 @@ namespace Azure.Communication.CallAutomation.Tests.Infrastructure
         public string TargetPhoneNumber => GetRecordedVariable(TargetNumber, options => options.IsSecret("+16041234567"));
 
         /// <summary>
+        /// Endpoint for the targetted PMA in string. If not set, default endpoint is used.
+        /// </summary>
+        public string PMAEndpoint => GetRecordedOptionalVariable(Endpoint, options => options.IsSecret("https://sanitized.com"));
+
+        /// <summary>
         /// Dispatcher endpoint for automated testing
         /// </summary>
         public string DispatcherEndpoint => GetRecordedOptionalVariable(dispatcherEndpoint, options => options.IsSecret("https://sanitized.skype.com"));
+
+        /// <summary>
+        /// Bot App Id for Dialog tests.
+        /// </summary>
+        public string BotAppId => GetRecordedOptionalVariable(botAppId, options => options.IsSecret("Sanitized"));
 
         /// <summary>
         /// ServiceBus string